pyproject.toml
[tool.poetry]
name = "streamlit-api-wortkshop"
version = "0.1.0"
description = ""
authors = ["Luiz Fernando <luizfsoliveira.lm@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.41.1"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.36"
plotly = "^5.24.1"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
